+++ docs/changelog.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Documentation in scn-tex format
 - README
 - Changelog
+- Add always generating for conclusion in implication
 
 ### Fixed
 - Fixed applying rules chain